Refactor bootloader config enums

+++ src/TMC9660Bootloader.cpp
@@ -79,10 +79,10 @@ bool TMC9660Bootloader::applyConfiguration(const BootloaderConfig &cfg) noexcept
   if (!setAddress(bootaddr::COMM_CONFIG))
     return false;
   uint32_t comm = 0;
-  comm |= (cfg.rs485.txen_pin & 0x3) << 5;          // BL_UART_TXEN
+  comm |= (static_cast<uint32_t>(cfg.rs485.txen_pin) & 0x3) << 5; // BL_UART_TXEN
   comm |= (cfg.spiComm.disable_spi ? 1u : 0u) << 1; // BL_DISABLE_SPI
-  comm |= (cfg.spiComm.boot_spi_iface & 0x1) << 2;  // BL_SPI_SELECT
-  comm |= (cfg.spiComm.spi0_sck_pin & 0x1) << 10;   // BL_SPI0_SCK
+  comm |= (static_cast<uint32_t>(cfg.spiComm.boot_spi_iface) & 0x1) << 2; // BL_SPI_SELECT
+  comm |= (static_cast<uint32_t>(cfg.spiComm.spi0_sck_pin) & 0x1) << 10;  // BL_SPI0_SCK
   if (!write32(comm))
     return false;
 
@@ -91,7 +91,7 @@ bool TMC9660Bootloader::applyConfiguration(const BootloaderConfig &cfg) noexcept
     return false;
   uint32_t flash = (cfg.spiFlash.enable_flash ? 1u : 0u);
   flash |= (static_cast<uint32_t>(cfg.spiFlash.cs_pin & 0x1F) << 3);
-  flash |= (static_cast<uint32_t>(cfg.spiFlash.freq_div & 0xF) << 8);
+  flash |= (static_cast<uint32_t>(cfg.spiFlash.freq_div) & 0xF) << 8;
   if (!write32(flash))
     return false;
 
@@ -99,10 +99,10 @@ bool TMC9660Bootloader::applyConfiguration(const BootloaderConfig &cfg) noexcept
   if (!setAddress(bootaddr::I2C_CONFIG))
     return false;
   uint32_t i2c = (cfg.i2c.enable_eeprom ? 1u : 0u);
-  i2c |= (static_cast<uint32_t>(cfg.i2c.sda_pin & 0x3) << 1);
-  i2c |= (static_cast<uint32_t>(cfg.i2c.scl_pin & 0x3) << 3);
+  i2c |= (static_cast<uint32_t>(cfg.i2c.sda_pin) & 0x3) << 1;
+  i2c |= (static_cast<uint32_t>(cfg.i2c.scl_pin) & 0x3) << 3;
   i2c |= (static_cast<uint32_t>(cfg.i2c.address_bits & 0x7) << 5);
-  i2c |= (static_cast<uint32_t>(cfg.i2c.freq_code & 0x7) << 8);
+  i2c |= (static_cast<uint32_t>(cfg.i2c.freq_code) & 0x7) << 8;
   if (!write32(i2c))
     return false;
 
@@ -122,13 +122,13 @@ bool TMC9660Bootloader::applyConfiguration(const BootloaderConfig &cfg) noexcept
   if (!setAddress(bootaddr::CLOCK_CONFIG))
     return false;
   uint32_t clk = 0;
-  clk |= (cfg.clock.use_external & 0x1) << 8;          // EXT_NOT_INT
-  clk |= (cfg.clock.xtal_drive & 0x7) << 9;            // XTAL_CFG
+  clk |= (static_cast<uint32_t>(cfg.clock.use_external) & 0x1) << 8; // EXT_NOT_INT
+  clk |= (static_cast<uint32_t>(cfg.clock.xtal_drive) & 0x7) << 9;   // XTAL_CFG
   clk |= (cfg.clock.xtal_boost ? 1u : 0u) << 12;       // XTAL_BOOST
-  clk |= (cfg.clock.ext_source_type & 0x1) << 13;      // EXT_NOT_XTAL
-  clk |= (cfg.clock.pll_selection & 0x3) << 16;        // PLL_OUT_SEL
+  clk |= (static_cast<uint32_t>(cfg.clock.ext_source_type) & 0x1) << 13; // EXT_NOT_XTAL
+  clk |= (static_cast<uint32_t>(cfg.clock.pll_selection) & 0x3) << 16;   // PLL_OUT_SEL
   clk |= (cfg.clock.rdiv & 0x1F) << 18;                // RDIV
-  clk |= (cfg.clock.sysclk_div & 0x3) << 23;           // SYS_CLK_DIV
+  clk |= (static_cast<uint32_t>(cfg.clock.sysclk_div) & 0x3) << 23;    // SYS_CLK_DIV
   if (!write32(clk))
     return false;
